ranking/requirements.txt
functions-framework==3.*
requests
python-dotenv
firebase-admin
datetime
google
